+++ multiplication.cpp
@@ -1,6 +1,6 @@
 #include "multiplication.hpp"
 
-int Multiplication::eval(int a, int b) {
+long long Multiplication::eval(long long a, long long b) {
   return a * b;
 }
